+++ src/sqlCVS/A_UpdateSchema.h
@@ -53,7 +53,7 @@ public:
 	 * @brief returns the id of the updated row
 	 */
 
-	virtual unsigned long ID() { return A_UPDATE_SCHEMA; }
+	virtual uint32_t ID() { return A_UPDATE_SCHEMA; }
 	
 	/**
 	 * @brief sets up the serialization
