+++ src/sqlCVS/R_GetRow.h
@@ -63,7 +63,7 @@ public:
 	/**
 	 * @brief  Returns the id of the request
 	 */
-	virtual unsigned long ID( ) { return R_GET_ROW; }
+	virtual uint32_t ID( ) { return R_GET_ROW; }
 
 	/**
 	 * @brief Sets up the serialization request
